ztp: OCPBUGS-23194: Add support for external node files
The siteConfig file size can grow quite large if we have a cluster with
multiple nodes that each include various networking configs.  This leads
to kustomize not being able to invoke the siteConfig plugin.
This change introduces a way to load the "nodes:" section of the
siteConfig from external files using "nodePath:".
The argument of "nodePath:" can point to a single file or a directory of
nodes and can also be mixed with other nodes from the "nodes:" section